+++ crates/server/src/bin/generate_types.rs
@@ -30,14 +30,12 @@ fn generate_types_content() -> String {
         server::routes::config::Environment::decl(),
         server::routes::task_attempts::CreateFollowUpAttempt::decl(),
         services::services::config::Config::decl(),
-        services::services::config::EnvironmentInfo::decl(),
         services::services::config::NotificationConfig::decl(),
         services::services::config::ThemeMode::decl(),
         services::services::config::EditorConfig::decl(),
         services::services::config::EditorType::decl(),
         services::services::config::GitHubConfig::decl(),
         services::services::config::SoundFile::decl(),
-        services::services::config::SoundConstants::decl(),
         services::services::git::GitBranch::decl(),
         services::services::git::BranchStatus::decl(),
         utils::diff::WorktreeDiff::decl(),
