+++ crates/services/src/services/config/versions/mod.rs
@@ -1,0 +1,2 @@
+pub(super) mod v1;
+pub(super) mod v2;
